Trying to remove more potential ambiguity from the file-viewer.
Refs #8220 but may not fix it yet.

* Simplifies the caching a lot by not pre-calculating the urls
* Doesn't return empty lists in case no extraction happened, re-extracts
or throws an error
