+++ src/olympia/files/decorators.py
@@ -15,7 +15,7 @@ from olympia.access import acl
 from olympia.addons.decorators import owner_or_unlisted_reviewer
 from olympia.amo.cache_nuggets import Token
 from olympia.files.models import File
-from olympia.files.templatetags.jinja_helpers import DiffHelper, FileViewer
+from olympia.files.file_viewer import DiffHelper, FileViewer
 
 
 log = olympia.core.logger.getLogger('z.addons')
