+++ src/olympia/files/templates/files/file.html
@@ -1,5 +1,7 @@
 <p>
-    <a href="{{ selected['url_serve'] }}">{{ _('Download {0}').format(selected['filename']) }}</a><br/>
+    <a href="{{ url('files.redirect', selected['id'], selected['short']) }}">
+        {{ _('Download {0}').format(selected['filename']) }}
+    </a><br/>
     {% if selected['msg'] %}<b class="error">{{ selected['msg'] }}</b><br/>{% endif %}
     {% trans version=selected['version'], size=selected['size']|filesizeformat(binary=True),
              sha256=selected['sha256'], mimetype=selected['mimetype'] %}
